extract list seriallizers from betaAPI package
