+++ src/internal/m365/collection/site/restore.go
@@ -3,7 +3,6 @@ package site
 import (
 	"context"
 	"errors"
-	"fmt"
 	"io"
 	"runtime/trace"
 
@@ -12,7 +11,6 @@ import (
 
 	"github.com/alcionai/corso/src/internal/common/dttm"
 	"github.com/alcionai/corso/src/internal/common/idname"
-	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/data"
 	"github.com/alcionai/corso/src/internal/diagnostics"
 	"github.com/alcionai/corso/src/internal/m365/collection/drive"
@@ -154,51 +152,10 @@ func restoreListItem(
 		return dii, clues.WrapWC(ctx, err, "reading backup data")
 	}
 
-	oldList, err := betaAPI.CreateListFromBytes(byteArray)
-	if err != nil {
-		return dii, clues.WrapWC(ctx, err, "creating item")
-	}
+	_ = listName
 
-	if name, ok := ptr.ValOK(oldList.GetDisplayName()); ok {
-		listName = name
-	}
-
-	var (
-		newName  = fmt.Sprintf("%s_%s", destName, listName)
-		newList  = betaAPI.ToListable(oldList, newName)
-		contents = make([]models.ListItemable, 0)
-	)
-
-	for _, itm := range oldList.GetItems() {
-		temp := betaAPI.CloneListItem(itm)
-		contents = append(contents, temp)
-	}
-
-	newList.SetItems(contents)
-
-	// Restore to List base to M365 back store
-	restoredList, err := service.Client().Sites().BySiteId(siteID).Lists().Post(ctx, newList, nil)
-	if err != nil {
-		return dii, graph.Wrap(ctx, err, "restoring list")
-	}
-
-	// Uploading of ListItems is conducted after the List is restored
-	// Reference: https://learn.microsoft.com/en-us/graph/api/listitem-create?view=graph-rest-1.0&tabs=http
-	if len(contents) > 0 {
-		for _, lItem := range contents {
-			_, err := service.Client().
-				Sites().
-				BySiteId(siteID).
-				Lists().
-				ByListId(ptr.Val(restoredList.GetId())).
-				Items().
-				Post(ctx, lItem, nil)
-			if err != nil {
-				return dii, graph.Wrap(ctx, err, "restoring list items").
-					With("restored_list_id", ptr.Val(restoredList.GetId()))
-			}
-		}
-	}
+	// TODO: use restore handler to create list and list items
+	restoredList := models.NewList()
 
 	dii.SharePoint = ListToSPInfo(restoredList, int64(len(byteArray)))
